GEODE-2741: Removes last of SharedPtr/SharedBase.
- Fixes Solaris compilation.
 - Removes template parameter packing.
 - Updates default set of linker libraries.
- Removes UnwrapSharedPtr and things of that nature.
- Removes TypeHelper
- Removes SharedPtr alias.
- Removed custom new/delete operators.
- Fixes std::cout usage.

+++ src/clicache/src/impl/AuthenticatedCache.cpp
@@ -111,7 +111,7 @@ namespace Apache
         {
           GC::KeepAlive(m_nativeptr);
         }
-        auto rootRegions = gcnew array<IRegion<TKey, TValue>^>( vrr.size( ) );
+        auto rootRegions = gcnew array<IRegion<TKey, TValue>^>( static_cast<int>(vrr.size( )) );
 
         for( System::Int32 index = 0; index < vrr.size( ); index++ )
         {
